requirements.txt
# Vision & ML
ultralytics==8.2.0
opencv-python==4.9.0.80
Pillow==10.3.0

# Multi-agent
crewai==0.28.0
langchain==0.1.20

# Google Cloud
google-cloud-storage==2.16.0
google-cloud-vision==3.7.2

# API & Backend
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1

# Dashboard
streamlit==1.35.0

# Utilities
pydantic==2.7.1
loguru==0.7.2
numpy==1.26.4

# Testing
pytest==8.2.0
pytest-cov==5.0.0
